requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==20.2.0
CacheControl==0.12.6
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
firebase-admin==4.4.0
google-api-core==1.22.3
google-api-python-client==1.12.3
google-auth==1.22.0
google-auth-httplib2==0.0.4
google-cloud-core==1.4.2
google-cloud-firestore==1.9.0
google-cloud-storage==1.31.2
google-crc32c==1.0.0
google-resumable-media==1.0.0
googleapis-common-protos==1.52.0
grpcio==1.32.0
httplib2==0.18.1
idna==2.10
msgpack==1.0.0
multidict==4.7.6
oauthlib==3.1.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PySocks==1.7.1
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
tweepy==3.9.0
uritemplate==3.0.1
urllib3==1.25.10
yarl==1.6.0
